Adding new Vorpal ponderomotive force data
When (a) UEDGE was run with the mod3.009.h5 density profile, (b) Vorpal was then run with that equilibrium profile, and we (c) computed the ponderomotive forces contained in VorpalDataOnUedgeGrid.nonuniform007.h5.  Then (d) UEDGE was run with the added ponderomotive forces (at 59%) to generate a perturbed density profile tmpnew_e0.590_i0.590.h5, and (e) Vorpal was rerun with that perturbed profile to compute new ponderomotive forces in this file, VorpalDataOnUedgeGrid.nonuniform009.h5.  (The computed forces look very similar to the old ones, in VorpalDataOnUedgeGrid.nonuniform007.h5, which suggests that convergence is not unlikely.  But that probably needs to be checked in more detail.)