Ponderomotive force data from Vorpal
For the low-density (1e16) case, this file VorpalDataOnUedgeGrid.nonuniform010.h5 contains ponderomotive force data computed by Vorpal on the equilibrium density profile.  This is similar to a case we've already looked at (the nonuniform001 file) and the UEDGE response will likely be very similar.  Nevertheless, I wanted to verify that each of the runs (low/medium/high density) were being done the same way in Vorpal and UEDGE.